recreate structure
